+++ app/models/__init__.py
@@ -1,5 +1,6 @@
 from .user import User
 from .alert import Alert
+from .alert_enrichment import AlertEnrichment
 from .customer import Customer
 from .source import Source
 from .case_note import CaseNote
@@ -20,6 +21,7 @@ __all__ = [
     "Case",
     "User",
     "Alert",
+    "AlertEnrichment",
     "Customer",
     "CustomerDetail",
     "CaseNote",